chore: move i18n.tsx into utils folder
